+++ NEWS.md
@@ -9,6 +9,7 @@
 * `headtail()`: Modified.  Internally remove `tbl_df` class (from `dplyr`) if it exists.  Added some checks with error messages.  Added suite of tests.
 * `lagratio()`: Modified.  Corrected incorrect explanation of `differences=` in help file.  Added `recursion=` and `direction=`.  Added suite of tests.
 * `mapvalues()`: Deleted.  Deleted the function from `FSA` but imported it from `plyr` and then exported it from `FSA` so that it would be available to fisheries users without having to load `plyr`.
+* `oddeven()`: Modified.  Added some checks with error messages.  Added suite of tests.
 
 # FSA 0.6.13 May15
 * Some miscellaneous reorganizations of files.